Conflicts:
	Projects/MacOSFramework/Framework.xcodeproj/project.pbxproj

Former-commit-id: c328e6bdb1ded264f40175856aaabd860eabf48c

+++ Projects/SceneViewer/Classes/SelectSceneScreen.cpp
@@ -47,8 +47,8 @@ void SelectSceneScreen::LoadResources()
     
     Rect screenRect = GetRect();
     Size2i screenSize = VirtualCoordinatesSystem::Instance()->GetVirtualScreenSize();
-    screenRect.dx = screenSize.dx;
-    screenRect.dy = screenSize.dy;
+    screenRect.dx = static_cast<float32>(screenSize.dx);
+    screenRect.dy = static_cast<float32>(screenSize.dy);
     SetRect(screenRect);
     const float32 buttonSize = 30.f;
